requirements.txt
datasets==1.17.0
datasets.egg==info
ftfy==6.0.3
hydra-core==1.1.1
omegaconf==2.1.1
pyarrow==3.0.0
Unidecode==1.3.2
